+++ package.json
@@ -42,10 +42,6 @@
 
 
     "build:css": "postcss src/assets/tailwind.css -o src/assets/main.css",
-
-
-    "build:css": "postcss src/assets/tailwind.css -o src/assets/main.css",
-
     "watch:css": "postcss src/assets/tailwind.css -o src/assets/main.css"
   },
   "eslintConfig": {
